Fix docker-compose migrations bug
- The api container was previously attempting to run migrations
while the postgres service wasn't ready. This was causing
an error and the notes database was always empty.
- The postgres-client package is now being installed in the api
container. This package contains a pg_ready command that is used
to check if the postgres service is ready to start accepting
connections. The polling technique is applied here.
- When the postgres service is ready to receive connections, the
migrations will be applied.